requirements.txt
google-generativeai
google-adk
requests
python-dotenv
numpy